+++ internal/client/cache.go
@@ -1,4 +1,4 @@
-package reader
+package client
 
 import (
 	"fmt"
