+++ internal/client/cache_test.go
@@ -1,8 +1,9 @@
-package reader
+package client
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestRead(t *testing.T) {
