+++ internal/client/client.go
@@ -1,4 +1,4 @@
-package reader
+package client
 
 import (
 	"errors"
