+++ internal/client/docker.go
@@ -1,4 +1,4 @@
-package reader
+package client
 
 // Docker represents a docker client.
 type Docker interface {
