+++ internal/client/docker_test.go
@@ -1,10 +1,11 @@
-package reader
+package client
 
 import (
 	"encoding/binary"
 	"errors"
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 type dockerMock struct {
